backend/service/requirements.txt
Django==5.0.2
django-admin==2.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
pika==1.3.2
pymongo==4.6.2
psycopg2==2.9.9

